Index * make superclass concrete * remove subclass
