+++ index/index.py
@@ -1,35 +1,46 @@
 """Module for a super class of a index used in a search engine."""
 
-from abc import ABC, abstractmethod
 import json
 import os
+from abc import abstractmethod
+
+from index import preprocess
 
 
-class Index(ABC):
-    """Abstract Base Class for a search index."""
+class Index:
+    """A basic inverted index."""
 
-    # TODO make the path check into a decorator
-    @abstractmethod
-    def load(self, path):
-        """Loads a stored conjunctive index."""
-        pass
+    def __init__(self):
+        self.index = {}
 
-    @abstractmethod
-    def save(self, path):
-        """Saves the conjunctive index into a file."""
-        pass
+    @staticmethod
+    def create_from(documents, words):
+        index = Index()
+        for name, desc in zip(documents, words):
 
-    @abstractmethod
-    def put(self, word_id, document: str):
+            desc = preprocess(desc)
+
+            for token in desc:
+                index.put(token, name)
+
+        return index
+
+
+    def put(self, word_id, document):
         """Puts a document into the inverted index at the word_id.
 
         :args
         word_id - the word for which the document will be added.
         document - the unique document name that will be added to the word in the index.
         """
-        pass
+        try:
+            self.index[word_id].add(document)
 
-    @abstractmethod
+        except KeyError:
+            documents = set()
+            documents.add(document)
+            self.index[word_id] = documents
+
     def get(self, word_id: int):
         """Gets a list of documents that belong to the given word.
 
@@ -39,9 +50,32 @@ class Index(ABC):
         :raises
         a KeyError if the word does not exist in the index.
         """
-        pass
+        try:
+            return self.index[word_id]
 
-    @abstractmethod
+        except KeyError:
+            raise KeyError(f'The word: {word_id}, does not exist in the index.')
+
+    def load(self, path):
+        if not os.path.isdir(path):
+            raise OSError(f'The path: {path} does not exist.')
+
+        file_name = 'conjunctive_index.json'
+        full_path = f'{path}/{file_name}'
+
+        with open(full_path, 'r') as file:
+            self.index = json.load(file)
+
+    def save(self, path):
+        if not os.path.isdir(path):
+            raise OSError(f'The path: {path} does not exist.')
+
+        file_name = 'conjunctive_index.json'
+        full_path = f'{path}/{file_name}'
+
+        with open(full_path, 'w') as file:
+            json.dump(self.index, file)
+
     def query(self, query):
         """Runs a query against the index.
 
@@ -51,3 +85,23 @@ class Index(ABC):
         :returns
         a list of document ids that match the query.
         """
+        query = preprocess(query)
+        found = []
+        for word in query:
+            try:
+                found.append(self.index[word])
+            except KeyError:
+                return []
+
+        return set.intersection(*map(set, found))
+
+    @abstractmethod
+    def query_top_k(self, query, k):
+        """Runs a query against the index and returns the top k documents.
+
+        :args
+        query - a query string.
+        k - an integer specifying the number of top ranked documents to return.
+
+        :returns
+        a list of length k of document ids that match the query."""
